PostgreSQL only allows indexes of up to 32 columns by default.

+++ lib/BIGSdb/CuratePage.pm
@@ -1115,13 +1115,15 @@ sub refresh_material_view {
 	}
 	eval {
 		$self->{'db'}->do("DROP INDEX i_mv$scheme_id\_1");
-		$self->{'db'}->do("DROP INDEX i_mv$scheme_id\_2");
 		$self->{'db'}->do("SELECT refresh_matview('mv_scheme_$scheme_id')");
 		$self->{'db'}->do("CREATE UNIQUE INDEX i_mv$scheme_id\_1 ON mv_scheme_$scheme_id ($pk)");
-		local $" = ',';
-		my $locus_string = "@$loci";
-		$locus_string =~ s/'/_PRIME_/g;
-		$self->{'db'}->do("CREATE UNIQUE INDEX i_mv$scheme_id\_2 ON mv_scheme_$scheme_id ($locus_string)");
+		if (@$loci <= 32){ #By default PostgreSQL will not allow indexes with more than 32 columns
+			local $" = ',';
+			my $locus_string = "@$loci";
+			$locus_string =~ s/'/_PRIME_/g;
+			$self->{'db'}->do("DROP INDEX i_mv$scheme_id\_2");
+			$self->{'db'}->do("CREATE UNIQUE INDEX i_mv$scheme_id\_2 ON mv_scheme_$scheme_id ($locus_string)");
+		}
 	};
 	$logger->error($@) if $@;
 	return;
